Add batch/rt example pages for nextjs
