+++ examples/nextjs/src/app/batch/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <h1>Coming soon!</h1>;
+}
